requirements.txt
pytest~=8.2.2
playwright
pytest-playwright
requests~=2.32.3
python-dotenv~=1.0.1
allure-pytest~=2.13.5
pytest-html~=4.1.1
python-dotenv~=1.0.1